requirements.txt
flask
flask-cors
mysql-connector-python
python-dotenv
numpy
rembg
pillow
requests
pandas
scikit-learn
